+++ src/main/java/dk/kb/license/facade/AuditLogModuleFacade.java
@@ -29,17 +29,17 @@ public class AuditLogModuleFacade {
      * AuditLogEntries in the list are sorted by modifiedTimeStart desc, so latest will come first in the list.     
      *      
      * @param modifiedTimeStart Will extract AuditLogEntries with modifiedTimeStart less than this value
-     * @param type Optional, list only AuditLogEntries of this type. Will list all types if type is null
+     * @param changeName list only AuditLogEntries of this type. Will list all types if type is null.
      * 
      * @return List<AuditLogEntryOutputDto> with a maximum of 100 elements in the list. 
      */
-    public static List<AuditLogEntryOutputDto> getAuditLogList(Long modifiedTimeStart, ObjectTypeEnumDto type) {                                  
+    public static List<AuditLogEntryOutputDto> getAuditLogOlderThanModifiedTime(Long modifiedTimeStart, ObjectTypeEnumDto changeName) {                                  
         return BaseModuleStorage.performStorageAction("getAuditLogList()", AuditLogModuleStorage.class, storage -> {
-         if (type == null) {
-             return ((AuditLogModuleStorage) storage).getAuditLogListAll(modifiedTimeStart);     
+         if (changeName == null) {
+             return ((AuditLogModuleStorage) storage).getAuditLogOlderThanModifiedTimeListAll(modifiedTimeStart);     
          }
          else {
-             return ((AuditLogModuleStorage) storage).getAuditLogListByType(modifiedTimeStart,type);
+             return ((AuditLogModuleStorage) storage).getAuditLogOlderThanModifiedTimeListByType(modifiedTimeStart,changeName);
          }        
         });
     }